requirements.txt
# Veri işleme
datasets==4.2.0
pandas==2.3.3
numpy==2.2.6

# Embedding ve Vector Store
sentence-transformers==5.1.1
chromadb==1.1.1
faiss-cpu==1.12.0

# LLM ve RAG
google-generativeai==0.8.3
langchain==0.3.27
langchain-core==0.3.79
langchain-google-genai==2.0.8
langchain-community==0.3.19

# Web Framework
Flask==3.1.2
python-dotenv==1.1.1

# Yardımcı
tqdm==4.67.1
